+truth.inject()
truth.inject() takes a proxy target & a delta comprised of the property path, action type (del, get, set) and potential new value and correctly applies the correct action to the object.

This has been spun off so that it can be used externally with ease to link proxies together to update data in two places simultaneously without having to write a bunch of event listeners to keep data in sync.