compiled pre/post ops at runtime
If a function is marked compiled (e.g. if(fn.compile)), then it will be passed the state & update functions directly and the return argument will be used in the pre/post op list.

This avoids the declaring the state store with truth and then adding pre/post ops after its inital state and update function have been generated.

truth.compile() may stilll be a bit awkward, but it is leaps and bounds bettter than before.